# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IIECS-101 Attendance & Invoice Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: batch
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Student not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}/id-card.pdf:
    get:
      operationId: downloadStudentIdCardPdf
      tags: [students]
      summary: Download the student's ID card PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ID card PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /students/by-email/{email}:
    get:
      operationId: getStudentByEmail
      tags: [students]
      summary: Get student by email
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
  /students/{id}/progress:
    get:
      operationId: getStudentProgress
      tags: [students]
      summary: Get student overall progress (attendance % + payment summary)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProgress"

  # ── Attendance ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, email, fullName, idNumber, batch, status, enrollmentDate, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        idNumber:
          type: string
        batch:
          type: string
        qrCodeData:
          type: ["string", "null"]
        idCardUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        cnic:
          type: ["string", "null"]
        status:
          type: string
        enrollmentDate:
          type: string
        createdAt:
          type: string
    StudentInput:
      type: object
      required: [email, fullName, idNumber, batch]
      properties:
        email:
          type: string
        fullName:
          type: string
        idNumber:
          type: string
        batch:
          type: string
        idCardUrl:
          type: string
        phone:
          type: string
        address:
          type: string
        cnic:
          type: string
    StudentUpdate:
      type: object
      properties:
        fullName:
          type: string
        idCardUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        cnic:
          type: ["string", "null"]
        status:
          type: string
    StudentProgress:
      type: object
      required: [studentId, fullName, overallAttendancePercentage, totalAttendanceRecords, totalPaid, totalPending]
      properties:
        studentId:
          type: string
        fullName:
          type: string
        overallAttendancePercentage:
          type: number
        totalAttendanceRecords:
          type: integer
        daysPresent:
          type: integer
        daysAbsent:
          type: integer
        daysLate:
          type: integer
        totalPaid:
          type: number
        totalPending:
          type: number

    # ── Attendance ───────────────────────────────────────────────────────────
...