# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IIECS-101 Attendance & Invoice Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  email: string;
  fullName: string;
  idNumber: string;
  batch: string;
  idCardUrl?: string;
  phone?: string;
  address?: string;
  cnic?: string;
}