# lib/api-zod/src/generated/types/studentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IIECS-101 Attendance & Invoice Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentUpdate {
  fullName?: string;
  /** @nullable */
  idCardUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  cnic?: string | null;
  status?: string;
}